Prevent builder from renaming node_modules directories.
If multiple directories are copied to the same output path by the builder,
in some cases it makes sense to ensure distinct directory names by adding
numeric suffixes to some of the directories.

In general, builder.generateFilename can get away with this renaming only
if the exact names of the directories are an implementation detail.

However, the code changed by this commit was altering the names of
node_modules directories whenever a package had both an Npm.depends and a
local node_modules directory.

Not only is it totally invalid to change the name of a node_modules
directory, but there is also no harm in copying the contents of multiple
node_modules directories into one final directory called node_modules.

Should fix #9738.